Create main.cpp
All the Lines which have SQL in it can be commented or removed.
This code is built store all information or data entered in a SQL table.
Use Xampp first. Then run Apache and MySQL. 
Open http://localhost/phpmyadmin/ on a website.
Create a table named employee and the columns according to the program.
Then execute the program.